+++ src/error.rs
@@ -2,69 +2,84 @@ use std::{fmt, io};
 
 use crate::MANIFEST_DIR;
 
-/// An error that occurred when getting manifest.
+/// Alias for a `Result` with the error type `find_crate::Error`.
+pub type Result<T, E = Error> = std::result::Result<T, E>;
+
+/// An error that occurred during during reading or parsing the manifest.
 #[derive(Debug)]
-pub enum Error {
-    /// The [`CARGO_MANIFEST_DIR`] environment variable not found.
-    ///
-    /// [`CARGO_MANIFEST_DIR`]: https://doc.rust-lang.org/cargo/reference/environment-variables.html#environment-variables-cargo-sets-for-crates
+pub struct Error(ErrorKind);
+
+// Hiding error variants from a library's public error type to prevent
+// dependency updates from becoming breaking changes.
+// We can add `is_*` methods that indicate the kind of error if needed, but
+// don't expose dependencies' types directly in the public API.
+/// An error that occurred during during reading or parsing the manifest.
+#[derive(Debug)]
+pub(crate) enum ErrorKind {
+    /// The `CARGO_MANIFEST_DIR` environment variable not found.
     NotFoundManifestDir,
 
     /// The manifest is invalid for the following reason.
     InvalidManifest(String),
 
-    /// The crate with the specified name not found. This error occurs only from [`find_crate`].
-    ///
-    /// [`find_crate`]: super::find_crate
-    NotFound,
-
-    /// An error occurred while to open or to read the manifest file.
+    /// An error that occurred during opening or reading the manifest file.
     Io(io::Error),
 
-    /// An error occurred while to parse the manifest file.
+    /// An error that occurred during parsing the manifest file.
     Toml(toml::de::Error),
+}
 
-    #[doc(hidden)]
-    __Nonexhaustive,
+impl Error {
+    pub(crate) fn new(e: impl Into<ErrorKind>) -> Self {
+        Error(e.into())
+    }
+
+    pub(crate) fn invalid_manifest(s: impl Into<String>) -> Self {
+        Error(ErrorKind::InvalidManifest(s.into()))
+    }
 }
 
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Error::NotFoundManifestDir => {
+        match &self.0 {
+            ErrorKind::NotFoundManifestDir => {
                 write!(f, "`{}` environment variable not found", MANIFEST_DIR)
             }
-            Error::InvalidManifest(reason) => {
+            ErrorKind::InvalidManifest(reason) => {
                 write!(f, "The manifest is invalid because: {}", reason)
             }
-            Error::NotFound => {
-                write!(f, "the crate with the specified name not found in dependencies")
+            ErrorKind::Io(e) => write!(f, "an error occurred while to open or to read: {}", e),
+            ErrorKind::Toml(e) => {
+                write!(f, "an error occurred while parsing the manifest file: {}", e)
             }
-            Error::Io(e) => write!(f, "an error occurred while to open or to read: {}", e),
-            Error::Toml(e) => write!(f, "an error occurred while parsing the manifest file: {}", e),
-            Error::__Nonexhaustive => unreachable!(),
         }
     }
 }
 
 impl std::error::Error for Error {
     fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        match self {
-            Error::Io(e) => Some(e),
-            Error::Toml(e) => Some(e),
+        match &self.0 {
+            ErrorKind::Io(e) => Some(e),
+            ErrorKind::Toml(e) => Some(e),
             _ => None,
         }
     }
 }
 
-impl From<io::Error> for Error {
-    fn from(e: io::Error) -> Self {
-        Error::Io(e)
+impl From<ErrorKind> for Error {
+    fn from(kind: ErrorKind) -> Self {
+        Error(kind)
     }
 }
 
-impl From<toml::de::Error> for Error {
+impl From<io::Error> for ErrorKind {
+    fn from(e: io::Error) -> Self {
+        ErrorKind::Io(e)
+    }
+}
+
+impl From<toml::de::Error> for ErrorKind {
     fn from(e: toml::de::Error) -> Self {
-        Error::Toml(e)
+        ErrorKind::Toml(e)
     }
 }