+++ CHANGELOG.md
@@ -10,6 +10,28 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- `find_crate` function is now returns `Result<Option<String>>` instead of `Result<Package>`. `Ok(None)` is equivalent to `Err(Error::NotFound)` in previous version.
+
+- `Manifest::find` method is now returns `Option<String>` instead of `Option<Package>`. If you want to get `Package`, use `Manifest::find_package` method instead.
+
+- Rename `Manifest::find2` method to `Manifest::find_package`.
+
+- `Manifest::from_toml` function is now takes `&str` instead of `toml::Table`.
+
+- Change `Error` from enum to struct.
+
+- Change `Package::version` field from `String` to `&str`.
+
+- Make `Dependencies` non-exhaustive.
+
+- Add `Result` type alias.
+
+- Remove `Package::is_original` method.
+
+- Change closures in argument position from type parameter to impl trait.
+
+- Documentation improvements.
+
 ## [0.6.3] - 2021-01-05
 
 - Exclude unneeded files from crates.io.
@@ -30,37 +52,37 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [0.5.0] - 2019-09-29
 
-- Made `Manifest::dependencies` and `Package::{name, varsion}` fields public.
+- Make `Manifest::dependencies` and `Package::{name, varsion}` fields public.
 
-- Added support for `target.cfg.dependencies`.
+- Add support for `target.cfg.dependencies`.
 
-- Added `Dependencies` enum to manage the kind of dependencies to be searched.
+- Add `Dependencies` enum to manage the kind of dependencies to be searched.
 
-- Removed `Manifest::lock()` and `ManifestLock`.
+- Remove `Manifest::lock` method and `ManifestLock`.
 
-- Removed some variant and field form `Error`.
+- Remove some variants and fields form `Error`.
 
-- Removed `DEFAULT_DEPENDENCIES`.
+- Remove `DEFAULT_DEPENDENCIES` constant.
 
 ## [0.4.0] - 2019-06-16
 
 - Transition to Rust 2018. With this change, the minimum required version will go up to Rust 1.31.
 
-- Updated minimum `toml` version to 0.5.0.
+- Update minimum `toml` version to 0.5.0.
 
 ## [0.3.0] - 2019-02-21
 
-- Removed version dependent behavior.
+- Remove version dependent behavior.
 
 - Documentation improvements.
 
 ## [0.2.0] - 2019-02-13
 
-- Supported Rust 1.15.
+- Support Rust 1.15.
 
 ## [0.1.2] - 2019-02-13
 
-- Implemented `PartialEq` and `Eq` for `Package`.
+- Implement `PartialEq` and `Eq` for `Package`.
 
 ## [0.1.1] - 2019-02-13